Problem: Can't template files
Solution: Start with the ability to template a file and do passthroughs

We are also careful to allow us a _common directory for common patterns
we wish to have available to us.

This is a first pass, I'm not sure how nicely it will play with sources
coming from all over. If jobs come from random places, will a _common be
a help or a hinderance? At least we can simply not include it, and do a
symlink instead, to find the requisite files...